refactor(preview-discovery): lift ClassGraph scan logic out of DiscoverPreviewsTask
Phase A2b of the contrib refactor — the actual scan that turns a class
directory + dependency-JAR bag into a `previews.json` manifest is now a
pure-JVM library function (`PreviewDiscovery.discover(input)`) sitting
in `:preview-discovery`. The Gradle plugin's `DiscoverPreviewsTask`
shrinks to a ~95-line adapter that resolves its Gradle-typed inputs to
`java.io.File` / `String` and routes the library's outcome through
`logger.warn` / `logger.lifecycle` / `GradleException`.

Non-Gradle consumers (Bazel rules, Amper tasks in
`yschimke/compose-ai-contrib`) call into `PreviewDiscovery` directly
from a `java -jar` invocation; A2c adds the CLI main.

Logger-agnostic contract: `PreviewDiscovery.Outcome` is a sealed class
returning either `Success(manifest, warnings, infoMessages)` or
`Failure(reason, diagnostics)`. The adapter routes those collections
to whatever logger its build system provides — no `Gradle.Logger`
reaches into the library.

Two sibling helpers moved along for the ride:

- `PreviewTargetInference` (ASM-driven `@Composable` call-target
  scoring) — no Gradle deps before the move; visibility bumped from
  `internal` to public so the adapter can address it without sharing a
  module.
- `DeviceDimensions` (preview device geometry resolver) — also no
  Gradle deps; lifts because `extractPreviewParams` calls into it from
  inside the scan loop.

The `:preview-discovery` artifact gains `api(libs.classgraph)` +
`api(libs.asm)` as transitives since both are part of the public API
surface (callers will see `ScanResult` / `ClassInfo` types in the
PreviewTargetInference signature).

Verified:
- `:gradle-plugin:preview-discovery:test` — `PreviewDataTest`,
  `DeviceDimensionsTest`, `PreviewTargetInferenceTest` all pass against
  the relocated copies.
- `:gradle-plugin:test` — unit tests still green.
- `:gradle-plugin:functionalTest --tests DiscoveryFunctionalTest` —
  end-to-end TestKit project still produces a conforming `previews.json`.
- `ktfmtCheckAll` clean.
